Renamed new quadbox to quadbox
