+++ packages/augur-ui/src/modules/account/components/augur-status.tsx
@@ -5,7 +5,7 @@ import Activity from 'modules/account/components/activity';
 import { AUGUR_STATUS_TITLE } from 'modules/common/constants';
 
 import Styles from 'modules/account/components/status.styles.less';
-import NewQuadBox from 'modules/portfolio/components/common/new-quad-box';
+import QuadBox from 'modules/portfolio/components/common/quad-box';
 import classNames from 'classnames';
 
 export interface AugurStatusProps {
@@ -15,7 +15,7 @@ export interface AugurStatusProps {
 const AugurStatus = ({
   hideHeader,
 }: AugurStatusProps) => (
-  <NewQuadBox
+  <QuadBox
     title={AUGUR_STATUS_TITLE}
     customClass={classNames(Styles.ShowHeaderOnMobile, {
       [Styles.HideHeader]: hideHeader,
