+++ packages/augur-ui/src/modules/create-market/saved-drafts.tsx
@@ -9,7 +9,7 @@ import { CancelTextButton } from 'modules/common/buttons';
 import { useAppStatusStore } from 'modules/app/store/app-status';
 import { createBigNumber } from 'utils/create-big-number';
 import { ZERO, ONE } from 'modules/common/constants';
-import NewQuadBox from 'modules/portfolio/components/common/new-quad-box';
+import QuadBox from 'modules/portfolio/components/common/quad-box';
 
 interface SavedDraftsProps {
   updatePage: Function;
@@ -71,7 +71,7 @@ const SavedDrafts = ({
   });
 
   return (
-    <NewQuadBox
+    <QuadBox
       title={'Saved drafts'}
       customClass={Styles.CustomQuadBoxStyles}
       // extraTitlePadding
