--- packages/augur-ui/src/modules/portfolio/components/common/box-header.styles.less
@@ -1,237 +1,0 @@
-@import (reference) '~assets/styles/shared';
-
-.BoxHeader {
-  > div:first-of-type {
-    align-items: center;
-    background: var(--color-dark-grey);
-    border-bottom: @size-1 solid var(--color-dark-grey);
-    display: flex;
-    flex-direction: row;
-    padding: 0 @size-10 0 @size-16;
-    max-height: @size-40;
-    min-height: @size-40;
-
-    > div:first-of-type {
-      .mono-11-bold;
-
-      line-height: @size-12;
-      margin-right: @size-16;
-      white-space: nowrap;
-      text-transform: uppercase;
-    }
-
-    > div:first-of-type > h1 {
-      .mono-11-bold;
-
-      margin-bottom: 0;
-    }
-
-    > div:last-of-type {
-      align-items: center;
-      display: flex;
-      flex: 1;
-      flex-direction: row;
-      justify-content: flex-end;
-    }
-  }
-
-  &.ExtraTitlePadding {
-    > div:first-of-type > div:first-of-type {
-      margin-left: @size-16;
-    }
-  }
-
-  &.NoBorders {
-    > div:first-of-type {
-      border-bottom: 0;
-      padding-bottom: 0;
-    }
-  }
-
-  @media @breakpoint-mobile {
-    &.Normal {
-      > div {
-        padding: 0;
-      }
-    }
-
-    &.NoBorders {
-      > div:first-of-type {
-        padding-left: @size-16;
-        min-height: @size-32;
-        max-height: @size-32;
-
-        > div,
-        > h1 {
-          .text-12-medium;
-        }
-      }
-    }
-  }
-}
-
-.BoxHeader > div:nth-of-type(2),
-.Title {
-  background: var(--color-table-header);
-  border-bottom: @size-1 solid var(--color-dark-grey);
-  align-items: center;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  padding: 0 @size-16 0;
-  height: @size-32;
-}
-
-.Title {
-  .text-12-medium;
-
-  background-color: var(--color-dark-grey);
-  text-transform: uppercase;
-}
-
-.RightContent,
-.MostRightContent,
-.BottomRightContent {
-  display: flex;
-  justify-content: flex-end;
-  padding: 0 @size-8;
-}
-
-.MostRightContent {
-  margin: @size-12 0;
-}
-
-.BottomRightContent {
-  margin-bottom: @size-12;
-}
-
-.BottomContent {
-  background: var(--color-table-header);
-  border: 1px solid var(--color-dark-grey);
-  margin-top: @size-16;
-}
-
-.NoTopMargin {
-  margin-top: 0;
-}
-
-.noBackground {
-  background: unset;
-  margin-top: 0;
-}
-
-.ShowOnMobile {
-  display: none;
-}
-
-.HideOnMobile {
-  display: block;
-}
-
-.Middle {
-  display: flex;
-  flex: 1;
-  flex-direction: column-reverse;
-}
-
-.isSwitched {
-  flex-direction: column;
-  order: -1;
-}
-
-.Toggle {
-  margin-left: @size-16;
-  display: flex;
-
-  > svg {
-    width: @size-12;
-    height: @size-22;
-
-    > path {
-      fill: var(--color-secondary-text);
-    }
-  }
-}
-
-@media @breakpoint-mobile {
-  .ShowOnMobile {
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-
-    > div:first-of-type {
-      display: flex;
-      justify-content: flex-end;
-      margin: @size-8 0;
-
-      > div:first-of-type {
-        flex: 1;
-      }
-
-      > div:nth-of-type(2) {
-        flex: 0;
-      }
-    }
-  }
-
-  .BottomRightContent {
-    margin-bottom: 0;
-  }
-
-  .HideOnMobile {
-    display: none;
-  }
-}
-
-:root[theme='BETTING'],
-:root[theme='SPORTS'] {
-  .BoxHeader {
-    > div:first-of-type {
-      background: transparent;
-      max-height: @size-56;
-      min-height: @size-56;
-      padding: 0 @size-12;
-
-      > div:first-of-type {
-        .text-16-bold;
-
-        margin-right: @size-8;
-        text-transform: capitalize;
-      }
-
-      > div:last-of-type {
-        div[class*='SearchBar'] {
-          > svg > g {
-            fill: var(--color-primary-text-50);
-          }
-        }
-
-        div[class*='isFocused'] {
-          max-width: @size-178;
-          margin-left: auto;
-          border: @size-1 solid var(--color-dark-text);
-
-          > input {
-            .text-14;
-
-            color: var(--color-primary-text);
-          }
-
-          > button > svg > path {
-            stroke: var(--color-primary-text);
-          }
-        }
-      }
-    }
-
-    > div:nth-of-type(2) {
-      padding: 0 @size-12;
-    }
-  }
-
-  .BoxHeader > div:nth-of-type(2),
-  .Title {
-    background: var(--color-module-background);
-    height: @size-40;
-  }
-}
