--- packages/augur-ui/src/modules/portfolio/components/common/new-quad-box.styles.less
@@ -1,333 +1,0 @@
-@import (reference) '~assets/styles/shared';
-
-.Search {
-  width: @size-34;
-  height: @size-36;
-  display: grid;
-  grid-template-columns: @size-28;
-  padding: @size-1 0;
-
-  > input {
-    .text-14;
-
-    display: none;
-    max-height: 100%;
-    background: var(--color-module-background);
-    border: 1px solid var(--color-secondary-action-outline);
-    border-radius: @border-radius-default;
-    color: var(--color-primary-text);
-    padding: @size-6 @size-34;
-    grid-column: 1 / span 3;
-    grid-row: 1;
-  }
-
-  > span {
-    display: none;
-    grid-column: 1;
-    grid-row: 1;
-    align-items: center;
-    justify-content: center;
-
-    > svg {
-      width: @size-14;
-      height: @size-14;
-      fill: var(--color-primary-text);
-    }
-  }
-
-  > button {
-    grid-column: 1;
-    grid-row: 1;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    > svg {
-      width: @size-16;
-      height: @size-16;
-      fill: var(--color-primary-text);
-
-      > path {
-        stroke: var(--color-inactive-text);
-      }
-    }
-  }
-
-  &.ShowInput {
-    grid-template-columns: @size-32 1fr @size-32;
-    width: 100%;
-
-    > input {
-      display: block;
-    }
-
-    > span {
-      display: flex;
-    }
-
-    > button {
-      top: @size-6;
-      grid-column: 3;
-      grid-row: 1;
-
-      > svg {
-        width: @size-10;
-        height: @size-10;
-      }
-    }
-  }
-
-  @media @breakpoint-mobile {
-    width: 100%;
-    height: @size-40;
-
-    > input {
-      display: block;
-    }
-
-    > span {
-      display: flex;
-    }
-
-    > button {
-      display: none;
-    }
-  }
-}
-
-.NewQuadBox {
-  color: var(--color-primary-text);
-  display: grid;
-  min-height: 50%;
-
-  &.HideToggle {
-    max-height: @size-43;
-    min-height: unset;
-    overflow-y: hidden;
-    flex: 0;
-  }
-
-  &.ExtendToggle {
-    min-height: calc(100% - @size-43);
-  }
-
-  .SearchAndFilters {
-    display: none;
-    grid-template-columns: auto;
-    grid-template-rows: @size-40;
-    grid-gap: @size-8;
-    padding: 0 0 @size-8 0;
-    align-items: center;
-
-    > div:first-of-type {
-    }
-
-    > button {
-      width: 100%;
-      height: @size-38;
-    }
-
-    &.Hide {
-      display: none;
-    }
-
-    &.TwoColumns {
-      grid-template-columns: 2fr 1fr;
-    }
-
-    @media @breakpoint-mobile {
-      display: grid;
-    }
-  }
-
-  .Container {
-    display: flex;
-    flex: 1;
-    flex-direction: column;
-    background: var(--color-module-background);
-    border: @size-1 solid var(--color-dark-grey);
-    border-radius: @border-radius-default;
-
-    .Header {
-      align-items: center;
-      background: var(--color-dark-grey);
-      border-bottom: @size-1 solid var(--color-dark-grey);
-      display: flex;
-      flex-direction: row;
-      padding: 0 @size-10 0 @size-16;
-      max-height: @size-40;
-      min-height: @size-40;
-
-      > span:first-of-type {
-        .mono-11-bold;
-
-        line-height: @size-12;
-        margin-right: @size-8;
-        white-space: nowrap;
-        text-transform: uppercase;
-      }
-
-      > div {
-        align-items: center;
-        display: flex;
-        flex: 1;
-        flex-direction: row;
-        justify-content: flex-end;
-
-        // search
-        > div:first-of-type {
-        }
-
-        // filters
-        > div:nth-of-type(2) {
-        }
-      }
-
-      .Count {
-        display: none;
-      }
-
-      &.Hide {
-        display: none;
-      }
-
-      .HideToggleButton {
-        > button:last-of-type {
-          display: none;
-        }
-      }
-    }
-
-    .Subheader {
-      background: var(--color-table-header);
-      border-bottom: 0.0625rem solid var(--color-dark-grey);
-      align-items: center;
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-start;
-      padding: 0 1rem 0;
-      height: 2rem;
-
-      &.Hide {
-        display: none;
-      }
-    }
-
-    .Content {
-      display: flex;
-      flex: 1;
-      flex-direction: column;
-
-      > span {
-        .text-12;
-
-        color: var(--color-inactive-text);
-      }
-    }
-  }
-
-  @media @breakpoint-mobile {
-    padding: @size-8;
-
-    .Container {
-      background: none;
-      border: none;
-      min-height: calc(100vh - 9.9rem);
-
-      .Header {
-        display: none;
-      }
-
-      .Subheader {
-        display: none;
-      }
-
-      .MarketBox {
-        border: @size-1 solid var(--color-dark-grey);
-        margin-top: @size-8;
-
-        > div[class*='MarketRow']:last-of-type {
-          border-bottom: none;
-        }
-      }
-    }
-  }
-}
-
-:root[theme='BETTING'],
-:root[theme='SPORTS'] {
-  .NewQuadBox {
-    .Container {
-      .Header {
-        background: transparent;
-        min-height: @size-56;
-        max-height: @size-56;
-        padding: 0 @size-12;
-
-        > span:first-of-type {
-          .text-16-bold;
-
-          text-transform: none;
-        }
-
-        .Count {
-          .text-12;
-
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          width: @size-19;
-          height: @size-21;
-          color: var(--color-secondary-text);
-          border: @size-1 solid var(--color-outline);
-          border-radius: @border-radius-default;
-          background: var(--color-module-background);
-          margin-right: auto;
-        }
-      }
-
-      .Subheader {
-        height: @size-40;
-        background: var(--color-module-background);
-        padding: 0 @size-12 0;
-
-        > div:first-of-type {
-          display: flex;
-          align-items: center;
-          width: 100%;
-
-          > span {
-            .text-14;
-
-            color: var(--color-secondary-text);
-          }
-
-          > span:nth-of-type(2) {
-            margin-left: auto;
-          }
-        }
-      }
-
-      .MarketBox span[class*="ArrowContainer"] {
-        display: none;
-      }
-    }
-  }
-
-  @media @breakpoint-mobile {
-    .Container {
-      background: var(--color-module-background);
-      border: @size-1 solid var(--color-dark-grey);
-      min-height: unset;
-    }
-
-    .Header {
-      display: none;
-    }
-
-    .MarketBox {
-      border: none;
-      margin-top: 0;
-    }
-  }
-}
