+++ packages/augur-ui/src/modules/portfolio/components/common/quad-box.tsx
@@ -1,100 +1,173 @@
-import React, { ReactNode } from 'react';
+import React, { ReactNode, useState } from 'react';
+import Styles from './quad-box.styles.less';
 import classNames from 'classnames';
-
-import BoxHeader from 'modules/portfolio/components/common/box-header';
-import { SearchSort } from 'modules/common/search-sort';
-
-import Styles from 'modules/portfolio/components/common/quad-box.styles.less';
+import { ToggleExtendButton } from 'modules/common/buttons';
+import { SearchIcon, XIcon } from 'modules/common/icons';
+import { SquareDropdown } from 'modules/common/selection';
+import { NameValuePair } from 'modules/portfolio/types';
+import SidebarNav, {
+  OPTIONTYPE,
+  TabsProps,
+} from 'modules/portfolio/components/common/sidebar-nav';
+import { useAppStatusStore } from 'modules/app/store/app-status';
+import { THEMES } from 'modules/common/constants';
 
 export interface QuadBoxProps {
-  bottomBarContent?: ReactNode;
-  bottomContent?: ReactNode;
-  bottomRightBarContent?: ReactNode;
+  title?: string;
+  headerComplement?: ReactNode;
+  subheader?: ReactNode;
   content?: ReactNode;
+  footer?: ReactNode;
   customClass?: string;
-  extend?: boolean;
-  extraTitlePadding?: boolean;
-  h1Title?: boolean;
-  hide?: boolean;
-  hideHeader?: boolean;
-  leftContent?: ReactNode;
-  noBackgroundBottom?: boolean;
-  noBorders?: boolean;
-  normalOnMobile?: boolean;
-  onSearchChange?: any;
-  rightContent?: ReactNode;
   search?: string;
-  showFilterSearch?: boolean | undefined;
-  showHeaderOnMobile?: boolean;
-  sortByOptions?: any;
-  sortByStyles?: object;
-  switchHeaders?: boolean;
-  title: string;
+  onSearchChange?: Function;
+  setSearch?: Function;
+  sortByOptions?: NameValuePair[];
+  updateDropdown?: Function;
+  tabs?: TabsProps[];
+  setSelectedTab?: Function;
   toggle?: Function;
-  updateDropdown?: any;
+  hide?: boolean;
+  extend?: boolean;
 }
 
-const BoxHeaderElement = (props: QuadBoxProps) => (
-  <BoxHeader
-    extraTitlePadding={props.extraTitlePadding}
-    title={props.title}
-    showHeaderOnMobile={props.showHeaderOnMobile}
-    normalOnMobile={props.normalOnMobile}
-    switchHeaders={props.switchHeaders}
-    noBorders={props.noBorders}
-    leftContent={props.leftContent}
-    rightContent={
-      (props.showFilterSearch && (
-        <SearchSort
-          sortByOptions={props.sortByOptions}
-          updateDropdown={props.updateDropdown}
-          sortByStyles={props.sortByStyles}
-          onChange={(value) => props.onSearchChange(value)}
-          checkBox={props.leftContent}
-        />
-      )) ||
-      props.rightContent
-    }
-    showToggle={Boolean(props.toggle)}
-    toggle={props.toggle}
-    hide={props.hide}
-    extend={props.extend}
-    bottomRightBarContent={props.bottomRightBarContent}
-    bottomBarContent={props.bottomBarContent}
-    noBackgroundBottom={props.noBackgroundBottom}
-    h1Title={props.h1Title}
-  />
-);
+const QuadBoxSearch = ({ search, setSearch }) => {
+  const [focus, setFocus] = useState(false);
 
-const QuadBox = (props: QuadBoxProps) => (
-  <div
-    className={classNames(Styles.Quad, {
-      [Styles.NoBorders]: props.noBorders,
-      [Styles.NormalOnMobile]: props.normalOnMobile,
-      [Styles.HideToggle]: props.hide,
-      [Styles.Extend]: props.extend,
-      [props.customClass]: props.customClass
-    })}
-  >
-    {!props.hideHeader &&
+  return setSearch ? (
+    <div
+      className={classNames(Styles.Search, {
+        [Styles.ShowInput]: focus,
+      })}
+    >
+      <input
+        type="text"
+        placeholder="Search"
+        value={search}
+        onChange={(event) => setSearch(event.target.value)}
+      />
+      <span>{SearchIcon}</span>
+      <button onClick={() => setFocus(!focus)}>
+        {focus ? XIcon : SearchIcon}
+      </button>
+    </div>
+  ) : null;
+};
+
+const QuadBoxSort = ({ sortByOptions, updateDropdown }) => {
+  return sortByOptions && updateDropdown ? (
+    <SquareDropdown
+      defaultValue={sortByOptions[0].value}
+      options={sortByOptions}
+      onChange={updateDropdown}
+    />
+  ) : null;
+};
+
+const QuadBoxMobileFilters = ({
+  sortByOptions,
+  updateDropdown,
+  tabs,
+  setSelectedTab,
+}) => {
+  const tabsFilters = setSelectedTab && tabs
+    ? {
+        type: OPTIONTYPE.RADIO,
+        action: setSelectedTab,
+        sectionTitle: 'Status',
+        selected: tabs[0].value,
+        options: tabs,
+      }
+    : null;
+
+  const sortFilters = updateDropdown && sortByOptions
+    ? {
+        type: OPTIONTYPE.RADIO,
+        action: updateDropdown,
+        sectionTitle: 'Order By',
+        selected: sortByOptions[0].value,
+        options: sortByOptions,
+      }
+    : null;
+
+  const filtersConfig = [tabsFilters, sortFilters].filter((filter) => !!filter);
+  const showMobileFilters = filtersConfig.length > 0;
+
+  return showMobileFilters ? (
+    <SidebarNav headerTitle="Filters" filters={filtersConfig} />
+  ) : null;
+};
+
+const QuadBox = ({
+  title,
+  headerComplement,
+  subheader,
+  content,
+  customClass,
+  search,
+  setSearch,
+  sortByOptions,
+  updateDropdown,
+  tabs,
+  setSelectedTab,
+  toggle,
+  hide,
+  extend,
+}: QuadBoxProps) => {
+  const { theme } = useAppStatusStore();
+  const isTrading = theme === THEMES.TRADING;
+  const nothingInTheHeader =
+    !title && !headerComplement && !setSearch && !updateDropdown;
+
+  return (
+    <div
+      className={classNames(Styles.NewQuadBox, {
+        [Styles.HideToggle]: hide,
+        [Styles.ExtendToggle]: extend,
+        [customClass]: customClass,
+      })}
+    >
       <div
-        className={classNames({ [Styles.HideOnMobile]: !props.normalOnMobile })}
-      >
-        <BoxHeaderElement {...props} switchHeaders={false} />
-      </div>
-    }
-    <div>
-      <div
-        className={classNames(Styles.ShowOnMobile, {
-          [Styles.Hide]: props.normalOnMobile,
+        className={classNames(Styles.SearchAndFilters, {
+          [Styles.Hide]: !setSearch && !updateDropdown && !setSelectedTab,
+          [Styles.TwoColumns]: setSearch && (updateDropdown || setSelectedTab),
         })}
       >
-        <BoxHeaderElement {...props} switchHeaders={props.switchHeaders} />
+        <QuadBoxSearch search={search} setSearch={setSearch} />
+        <QuadBoxMobileFilters
+          sortByOptions={sortByOptions}
+          updateDropdown={updateDropdown}
+          tabs={tabs}
+          setSelectedTab={setSelectedTab}
+        />
       </div>
-      {props.content ? props.content : null}
-      {props.bottomContent ? props.bottomContent : null}
+      <div className={Styles.Container}>
+        <div
+          className={classNames(Styles.Header, {
+            [Styles.Hide]: nothingInTheHeader,
+          })}
+        >
+          <span>{title}</span>
+          <div className={classNames({
+            [Styles.HideToggleButton]: !toggle,
+          })}>
+            {headerComplement && headerComplement}
+            {isTrading && <QuadBoxSort sortByOptions={sortByOptions} updateDropdown={updateDropdown} />}
+            <QuadBoxSearch search={search} setSearch={setSearch} />
+            <ToggleExtendButton toggle={toggle} />
+          </div>
+        </div>
+        <div
+          className={classNames(Styles.Subheader, {
+            [Styles.Hide]: !subheader,
+          })}
+        >
+          {subheader && subheader}
+        </div>
+        <div className={Styles.Content}>{content}</div>
+      </div>
     </div>
-  </div>
-);
+  );
+};
 
 export default QuadBox;
